+++ udata/core/dataset/api.py
@@ -54,6 +54,7 @@ from .api_fields import (
     license_fields,
     resource_fields,
     resource_type_fields,
+    upload_community_fields,
     upload_fields,
 )
 from .constants import RESOURCE_TYPES, UPDATE_FREQUENCIES
@@ -61,7 +62,12 @@ from .exceptions import (
     SchemasCacheUnavailableException,
     SchemasCatalogNotFoundException,
 )
-from .forms import CommunityResourceForm, DatasetForm, ResourceForm, ResourcesListForm
+from .forms import (
+    CommunityResourceForm,
+    DatasetForm,
+    ResourceFormWithoutId,
+    ResourcesListForm,
+)
 from .models import (
     Checksum,
     CommunityResource,
@@ -108,6 +114,7 @@ class DatasetApiParser(ModelApiParser):
         self.parser.add_argument("schema", type=str, location="args")
         self.parser.add_argument("schema_version", type=str, location="args")
         self.parser.add_argument("topic", type=str, location="args")
+        self.parser.add_argument("credit", type=str, location="args")
         self.parser.add_argument("dataservice", type=str, location="args")
 
     @staticmethod
@@ -377,8 +384,9 @@ class ResourcesAPI(API):
     def post(self, dataset):
         """Create a new resource for a given dataset"""
         ResourceEditPermission(dataset).test()
-        form = api.validate(ResourceForm)
+        form = api.validate(ResourceFormWithoutId)
         resource = Resource()
+
         if form._fields.get("filetype").data != "remote":
             api.abort(400, "This endpoint only supports remote resources")
         form.populate_obj(resource)
@@ -394,7 +402,21 @@ class ResourcesAPI(API):
     def put(self, dataset):
         """Reorder resources"""
         ResourceEditPermission(dataset).test()
-        data = {"resources": request.json}
+        resources = request.json
+        if len(dataset.resources) != len(resources):
+            api.abort(
+                400,
+                f"All resources must be reordered, you provided {len(resources)} "
+                f"out of {len(dataset.resources)}",
+            )
+        if set(r["id"] if isinstance(r, dict) else r for r in resources) != set(
+            str(r.id) for r in dataset.resources
+        ):
+            api.abort(
+                400,
+                f"Resource ids must match existing ones in dataset, ie: {set(str(r.id) for r in dataset.resources)}",
+            )
+        data = {"resources": resources}
         form = ResourcesListForm.from_json(
             data, obj=dataset, instance=dataset, meta={"csrf": False}
         )
@@ -451,7 +473,7 @@ class UploadNewCommunityResources(UploadMixin, API):
         responses={415: "Incorrect file content type", 400: "Upload error"},
     )
     @api.expect(upload_parser)
-    @api.marshal_with(upload_fields, code=201)
+    @api.marshal_with(upload_community_fields, code=201)
     def post(self, dataset):
         """Upload a new community resource"""
         infos = self.handle_upload(dataset)
@@ -510,7 +532,7 @@ class ReuploadCommunityResource(ResourceMixin, UploadMixin, API):
         "upload_community_resource",
         responses={415: "Incorrect file content type", 400: "Upload error"},
     )
-    @api.marshal_with(upload_fields)
+    @api.marshal_with(upload_community_fields)
     def post(self, community):
         """Update the file related to a given community resource"""
         ResourceEditPermission(community).test()
@@ -543,10 +565,19 @@ class ResourceAPI(ResourceMixin, API):
         """Update a given resource on a given dataset"""
         ResourceEditPermission(dataset).test()
         resource = self.get_resource_or_404(dataset, rid)
-        form = api.validate(ResourceForm, resource)
+        form = api.validate(ResourceFormWithoutId, resource)
+
+        # ensure filetype is not modified after creation
+        if (
+            form._fields.get("filetype").data
+            and form._fields.get("filetype").data != resource.filetype
+        ):
+            abort(400, "Cannot modify filetype after creation")
+
         # ensure API client does not override url on self-hosted resources
         if resource.filetype == "file":
             form._fields.get("url").data = resource.url
+
         # populate_obj populates existing resource object with the content of the form.
         # update_resource saves the updated resource dict to the database
         # the additional dataset.save is required as we update the last_modified date.