+++ udata/core/organization/api_fields.py
@@ -82,7 +82,9 @@ member_user_with_email_fields = api.inherit(
             readonly=True,
         ),
         "last_login_at": fields.Raw(
-            attribute=lambda o: o.last_login_at if check_can_access_user_private_info() else None,
+            attribute=lambda o: o.current_login_at
+            if check_can_access_user_private_info()
+            else None,
             description="The user last connection date (only present on show organization endpoint if the current user is member of the organization: admin or editor)",
             readonly=True,
         ),